moved `MediaPlaceholder` into top-level `ui` folder
- this is a re-usable UI component across various cards so it makes sense to keep it at the general UI components layer
